front-end/requirements.txt
streamlit>=1.36.0
requests>=2.31.0
pydub>=0.25.1
websocket-client>=1.8.0
